Project/email_service/requirements.txt
Django
django-filter
django-restframework
djangorestframework
djangorestframework-simplejwt
drf-yasg
markdown
pika
celery
redis
sendgrid